requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
tensorflow==2.20.0
numpy==1.26.4
pillow==10.2.0
gunicorn==22.0.0
